requirements.txt
aiogram==3.20.0.post0
aiohttp==3.11.18
aiosqlite==0.21.0
asyncpg==0.30.0
python-dotenv==1.1.0
SQLAlchemy==2.0.40
Telethon==1.40.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844 